+++ CHANGES.md
@@ -148,6 +148,8 @@ To be released.
     class.  [[#3071]]
  -  (@planetarium/account-web3-secret-storage) Added `Web3KeyMetadata`
     interface.  [[#3084]]
+ -  (Libplanet.Tools) Added `-v`/`--validator-key` option to
+    `planet block generate-genesis` command.  [[#3088]]
 
 ### Behavioral changes
 
@@ -265,6 +267,7 @@ To be released.
 [#3079]: https://github.com/planetarium/libplanet/pull/3079
 [#3080]: https://github.com/planetarium/libplanet/pull/3080
 [#3084]: https://github.com/planetarium/libplanet/pull/3084
+[#3088]: https://github.com/planetarium/libplanet/pull/3088
 
 
 Version 0.53.4